Code Refactoring Implemented MVC model
