+++ backend/app.js
@@ -1,0 +1,26 @@
+import express from "express";
+import dotenv from "dotenv";
+import cors from "cors";
+import cookieParser from "cookie-parser";
+
+import connectDB from "./utils/database.js";
+import userRoutes from "./routes/userRoutes.js";
+import productRoutes from "./routes/productRoutes.js";
+
+dotenv.config();
+const app = express();
+
+connectDB();
+
+// Middleware
+app.use(express.json());
+app.use(express.urlencoded({ extended: false }));
+app.use(cors({ origin: "http://localhost:5173", credentials: true }));
+app.use(cookieParser());
+app.use("/uploads", express.static("uploads"));
+
+// Routes
+app.use("/api/users", userRoutes);
+app.use("/api/products", productRoutes);
+
+export default app;
