+++ backend/server.js
@@ -1,0 +1,9 @@
+import app from "./app.js";
+import dotenv from "dotenv";
+
+dotenv.config();
+const PORT = process.env.PORT || 5000;
+
+app.listen(PORT, () => {
+  console.log(`Server running on http://localhost:${PORT}`);
+});
