+++ frontend/src/components/Signup.jsx
@@ -21,7 +21,7 @@ const Signup = () => {
       const userData = { name: name, email: email, password: password };
 
       axios
-        .post("http://localhost:3000/register", userData, {
+        .post("http://localhost:3000/api/users/register", userData, {
           withCredentials: true,
         })
         .then((response) => {
